third_party/WebKit/LayoutTests/svg/hittest/singular-transform-6.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<style>
body { margin: 0; }
</style>
<svg width="400" height="100" viewBox="0 0 0 0">
  <g class="p">
    <rect class="a" width="200" height="100"/>
    <rect class="b" width="100" height="100" fill="green"/>
  </g>
</svg>
<script>
if (window.testRunner)
  eventSender.mouseMoveTo(150, 50);

shouldBeNull("document.querySelector('.p:hover .a')");
</script>
