third_party/blink/web_tests/platform/generic/svg/hittest/singular-transform-6-expected.txt

FAIL document.querySelector('.p:hover .a') should be null. Was [object SVGRectElement].
PASS successfullyParsed is true

TEST COMPLETE
